+++ lib/user/user_data.dart
@@ -4,6 +4,8 @@ class UserData {
   int level;
   int expPoints;
   String? username;
+  bool canClaimDailyReward;
+  DateTime? lastDailyClaim;
 
   // constructor
   UserData({
@@ -11,6 +13,8 @@ class UserData {
     this.pfpBase64,
     this.level = 1,
     this.expPoints = 0,
+    this.canClaimDailyReward = true,
+    this.lastDailyClaim,
     String? username,
   }) : username = username ?? uid; // Default username is the UID
 }